pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "droidz-installer"
version = "0.1.0"
description = "Instruction-only multi-platform installer for Droidz"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "PyYAML>=6.0",
]

[tool.setuptools.packages.find]
include = ["droidz_installer"]

[project.scripts]
droidz-install = "droidz_installer.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
